requirements.txt
torch>=2.0.0
transformers>=4.30.0
safetensors>=0.3.1
faiss-cpu>=1.7.4  # Use faiss-gpu for GPU support
numpy>=1.24.0
tqdm>=4.65.0
setuptools>=65.0.0
wheel>=0.40.0
scikit-learn